+++ app/build.gradle
@@ -17,7 +17,7 @@ android {
         minSdk 26
         targetSdk 35
         versionCode 20
-        versionName "0.22.3"
+        versionName "0.22.4"
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
